Renamed timetypes to time_types to follow rust naming conventions.
