+++ ntp-proto/src/lib.rs
@@ -1,7 +1,7 @@
 mod clock;
 mod packet;
-mod timetypes;
+mod time_types;
 
 pub use clock::{NtpClock, SystemClock};
 pub use packet::NtpHeader;
-pub use timetypes::{NtpDuration, NtpTimestamp};
+pub use time_types::{NtpDuration, NtpTimestamp};
